Move disarmed event
